pyproject.toml
[project]
name = "gsssp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "colorama==0.4.6",
    "contourpy==1.1.1",
    "cycler==0.12.1",
    "fonttools==4.57.0",
    "importlib-resources==6.4.5",
    "keras-cv>=0.9.0",
    "kiwisolver==1.4.7",
    "matplotlib==3.7.5",
    "numpy==1.24.3",
    "opencv-python==4.11.0.86",
    "packaging==25.0",
    "pillow==10.4.0",
    "pyparsing==3.1.4",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "tensorflow[and-cuda]==2.16.1",
    "tqdm==4.67.1",
    "zipp==3.20.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
